Car authorization added

+++ CarRental.Application/Functions/Cars/Commands/DeleteCar/DeleteCarCommandHandler.cs
@@ -1,6 +1,10 @@
-﻿using CarRental.Application.Contracts.Persistence;
+﻿using CarRental.Application.Authorization.Common;
+using CarRental.Application.Contracts.Persistence;
+using CarRental.Application.Functions.CarAddresses.Exceptions;
 using CarRental.Application.Functions.RentalCalculator.Exceptions;
+using CarRental.Domain.Entities;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -12,10 +16,15 @@ namespace CarRental.Application.Functions.Cars.Commands.DeleteCar
     public class DeleteCarCommandHandler : IRequestHandler<DeleteCarCommand, Unit>
     {
         private readonly ICarRepository _carRepository;
+        private readonly ICarAddressRepository _carAddressRepository;
+        private readonly IAuthorizationService _authorizationService;
 
-        public DeleteCarCommandHandler(ICarRepository carRepository)
+        public DeleteCarCommandHandler(ICarRepository carRepository, IAuthorizationService authorizationService, 
+            ICarAddressRepository carAddressRepository)
         {
             _carRepository = carRepository;
+            _authorizationService = authorizationService;
+            _carAddressRepository = carAddressRepository;
         }
         public async Task<Unit> Handle(DeleteCarCommand request, CancellationToken cancellationToken)
         {
@@ -24,6 +33,16 @@ namespace CarRental.Application.Functions.Cars.Commands.DeleteCar
             if (car == null)
             {
                 throw new CarNotFoundException();
+            }
+
+            var carAddress = await _carAddressRepository.GetByIdAsync(car.CarAddressId);
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(request.User, carAddress,
+                new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
             }
 
             await _carRepository.DeleteAsync(car);
